+++ src/components/ssc-exercise/ClearLocalStorage.tsx
@@ -1,14 +1,14 @@
 'use client';
 
-import useRemoveLocalStorageItem from '@/hooks/useRemoveLocalStorageItem'; // Adjust path as needed
 import {useEffect} from 'react';
+import {useSSCChaptersHandler} from '@/hooks/useSSCChaptersHandler';
 
 const ClearLocalStorage = () => {
-  const removeItem = useRemoveLocalStorageItem('chapterDone');
+  const {removeLocalStorageItem} = useSSCChaptersHandler();
 
   useEffect(() => {
-    removeItem();
-  }, [removeItem]);
+    removeLocalStorageItem();
+  }, []);
 
   return null;
 };
